+++ packages/security/src/node/access/jexl-engine-provider.ts
@@ -1,0 +1,18 @@
+import { Component } from '@malagu/core';
+import { JexlEngineProvider } from './access-protocol';
+import { Jexl } from 'jexl';
+
+@Component(JexlEngineProvider)
+export class JexlEngineProviderImpl implements JexlEngineProvider<any> {
+
+    protected jexlEngine: any;
+
+    provide(): any {
+
+        if (!this.jexlEngine) {
+            this.jexlEngine = new Jexl();
+        }
+        return this.jexlEngine;
+    }
+
+}